% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_US_cities}
\alias{fit_US_cities}
\title{Fitted distributions to the US Census 2000 city size distribution.}
\format{
A data frame with 52 rows and 7 variables:
\describe{
  \item{dist}{distribution}
  \item{components}{number of components}
  \item{prior}{list of prior weights for the individual distribution components of FMM}
  \item{coefficients}{list of coefficients for the distributions}
  \item{np}{Number of paramters}
  \item{n}{Number of observations}
  \item{convergence}{Logical indicating whether the fitting procedure converged}
}
}
\source{
\url{http://doi.org/10.3886/E113328V1}
}
\usage{
fit_US_cities
}
\description{
A dataset containing 52 distribution fits to the US Census 2000 city size distributions
}
\keyword{datasets}
